+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Changes
 
-* `find_response()` now also works for objects from package *marginaleffects*.
+* `find_response()`, `find_predictors()`, and `find_variables()` now also work
+  for objects from package *marginaleffects*.
 
 # insight 1.5.2